Load protocol from file
